+++ server/changelog.md
@@ -1,2 +1,3 @@
 - Fixed "Incorrect behavior on resuming AudioSender". ([#501](https://github.com/plasmoapp/plasmo-voice/issues/501))
 - Added sanity bounds checks for collections decoding. ([#505](https://github.com/plasmoapp/plasmo-voice/issues/505))
+- Embedded player info request retry logic to improve connection reliability. ([#466](https://github.com/plasmoapp/plasmo-voice/issues/466))